[SOIN] Extrais un composant pour les cartes des items
