+++ lib-catalogue/src/Catalogue.svelte
@@ -5,13 +5,10 @@
   import FiltreTypologieEtFormat from "./FiltreTypologieEtFormat.svelte";
   import FiltreSource from "./FiltreSource.svelte";
   import FiltreTheme from "./FiltreTheme.svelte";
-  import { type ItemCyber, Typologie } from "./Catalogue.types";
   import { recherches } from "./stores/recherches.store";
+  import CarteItem from "./CarteItem.svelte";
 
   const reinitialiseFiltres = () => recherches.reinitialise();
-
-  const libelleBadge = (item: ItemCyber) =>
-    item.typologie === Typologie.SERVICE ? "Service" : item.format;
 </script>
 
 <div class="recherche">
@@ -57,29 +54,7 @@
 
   <div class="liste">
     {#each $catalogueFiltre.resultats as itemCyber}
-      <div class="carte {itemCyber.typologie}">
-        <figure>
-          <img
-            src="/assets/images/illustrations-services/{itemCyber.illustration}"
-            alt="Illustration du service"
-          />
-          <figcaption>{libelleBadge(itemCyber)}</figcaption>
-        </figure>
-        <div class="contenu">
-          <h3>{@html itemCyber.nom}</h3>
-          <span>{@html itemCyber.description}</span>
-          <a href={itemCyber.lienInterne}>
-            En savoir plus
-            <img
-              src="/assets/images/icone-fleche-droite.svg"
-              alt="En savoir plus"
-            />
-          </a>
-          <div class="labels">
-            {#each itemCyber.sources as source}<span>{source}</span>{/each}
-          </div>
-        </div>
-      </div>
+      <CarteItem {itemCyber} />
     {/each}
   </div>
 </div>
